+++ oc1_source/perturb.c
@@ -24,14 +24,14 @@ extern int no_of_coeffs;
 extern int no_of_dimensions,no_of_categories;
 extern int *left_count,*right_count;
 extern int coeff_modified;
-extern float *coeff_array;
-extern float *modified_coeff_array;
+extern double *coeff_array;
+extern double *modified_coeff_array;
 extern struct unidim *candidates;
 extern double *temp_val; /*Work area */
 
-float mygamma;
-float compute_impurity();
-float myabs(),myrandom();
+double mygamma;
+double compute_impurity();
+double myabs(),myrandom();
 double exp();
 
 
@@ -55,15 +55,15 @@ double exp();
 /* Important Variables used : gamma and lambda : see CART book, chapter */
 /*                            5 for a description.                      */
 /************************************************************************/
-float cart_perturb(cur_points,cur_no_of_points, cur_coeff,cur_error)
+double cart_perturb(cur_points,cur_no_of_points, cur_coeff,cur_error)
      POINT **cur_points;
-     float cur_error;
+     double cur_error;
      int cur_no_of_points,cur_coeff;
 {
   int i,j,bestsplit,no_of_eff_points;
-  float d_dim_error;
-  float x,linear_split();
-  float lambda, best_lambda,best_mygamma,best_impurity;
+  double d_dim_error;
+  double x,linear_split();
+  double lambda, best_lambda,best_mygamma,best_impurity;
   
   for (i=1;i<=no_of_coeffs;i++)
     modified_coeff_array[i] = coeff_array[i];
@@ -81,7 +81,7 @@ float cart_perturb(cur_points,cur_no_of_points, cur_coeff,cur_error)
 	    {
 	      no_of_eff_points++;
 	      candidates[no_of_eff_points].cat = cur_points[i]->category;
-	      x = (float)(cur_points[i]->val/(cur_points[i]->dimension[cur_coeff]
+	      x = (double)(cur_points[i]->val/(cur_points[i]->dimension[cur_coeff]
 					      + mygamma));
 	      candidates[no_of_eff_points].value = x;
 	    }
@@ -133,15 +133,15 @@ float cart_perturb(cur_points,cur_no_of_points, cur_coeff,cur_error)
 /* Important Variables used :
 /* Remarks :
 /************************************************************************/
-float cart_perturb_constant(cur_points,cur_no_of_points,cur_error)
+double cart_perturb_constant(cur_points,cur_no_of_points,cur_error)
      POINT **cur_points;
-     float cur_error;
+     double cur_error;
      int cur_no_of_points;
 {
   int i,j,bestsplit,no_of_eff_points;
-  float d_dim_error;
-  float x,linear_split();
-  float lambda;
+  double d_dim_error;
+  double x,linear_split();
+  double lambda;
   
   for (i=1;i<=no_of_coeffs;i++)
     modified_coeff_array[i] = coeff_array[i];
@@ -152,7 +152,7 @@ float cart_perturb_constant(cur_points,cur_no_of_points,cur_error)
     {
       no_of_eff_points++;
       candidates[no_of_eff_points].cat = cur_points[i]->category;
-      candidates[no_of_eff_points].value = (float)cur_points[i]->val;
+      candidates[no_of_eff_points].value = (double)cur_points[i]->val;
     }
   
   lambda = linear_split(no_of_eff_points);
@@ -204,17 +204,17 @@ float cart_perturb_constant(cur_points,cur_no_of_points,cur_error)
 /* Remarks :	For a detailed description of the perturbation 		*/
 /*		algorithm, see Murthy et al's paper in AAAI-93.		*/
 /************************************************************************/
-float suggest_perturbation(cur_points,cur_no_of_points,cur_coeff,cur_error)
+double suggest_perturbation(cur_points,cur_no_of_points,cur_coeff,cur_error)
      POINT **cur_points;
-     float cur_error;
+     double cur_error;
      int cur_no_of_points,cur_coeff;
 {
   extern int no_of_stagnant_perturbations;
   int i,j,lpt,rpt,bestsplit,no_of_eff_points=0;
-  float d_dim_error;
-  float suggest_perturbation();
-  float x,linear_split();
-  float newval,changeinval;
+  double d_dim_error;
+  double suggest_perturbation();
+  double x,linear_split();
+  double newval,changeinval;
 
   for (i=1;i<=no_of_coeffs;i++)
      modified_coeff_array[i] = coeff_array[i];
@@ -227,7 +227,7 @@ float suggest_perturbation(cur_points,cur_no_of_points,cur_coeff,cur_error)
 	no_of_eff_points++;
 	candidates[no_of_eff_points].cat = cur_points[i]->category;
 	candidates[no_of_eff_points].value =
-	  coeff_array[no_of_coeffs] - (float)cur_points[i]->val;
+	  coeff_array[no_of_coeffs] - (double)cur_points[i]->val;
       }
   else 
     for (i=1;i<=cur_no_of_points;i++)
@@ -235,7 +235,7 @@ float suggest_perturbation(cur_points,cur_no_of_points,cur_coeff,cur_error)
 	{
 	  no_of_eff_points++;
 	  candidates[no_of_eff_points].cat = cur_points[i]->category;
-	  x = (float)(cur_points[i]->val/cur_points[i]->dimension[cur_coeff]);
+	  x = (double)(cur_points[i]->val/cur_points[i]->dimension[cur_coeff]);
 	  candidates[no_of_eff_points].value = coeff_array[cur_coeff] - x;
 	}
 
@@ -301,17 +301,17 @@ float suggest_perturbation(cur_points,cur_no_of_points,cur_coeff,cur_error)
 /*			free_vector (util.c)				*/
 /* Is called by modules : oblique_split (mktree.c)			*/
 /************************************************************************/
-float perturb_randomly(cur_points,cur_no_of_points,cur_error,cur_label)
+double perturb_randomly(cur_points,cur_no_of_points,cur_error,cur_label)
      POINT **cur_points;
-     float cur_error;
+     double cur_error;
      int cur_no_of_points;
      char *cur_label;
 {
   extern int no_of_stagnant_perturbations;
   int i,j,no_of_eff_points=0;
-  float d_dim_error;
-  float *rvector;
-  float alpha,linear_split();
+  double d_dim_error;
+  double *rvector;
+  double alpha,linear_split();
   
   for (i=1;i<=no_of_coeffs;i++)
     modified_coeff_array[i] = coeff_array[i];
@@ -332,7 +332,7 @@ float perturb_randomly(cur_points,cur_no_of_points,cur_error,cur_label)
 	  no_of_eff_points++;
 	  candidates[no_of_eff_points].cat = cur_points[i]->category;
 	  candidates[no_of_eff_points].value = 
-	    (float)(-1.0 * (cur_points[i]->val / temp_val[i]));
+	    (double)(-1.0 * (cur_points[i]->val / temp_val[i]));
 	}
     }
   
@@ -395,7 +395,7 @@ float perturb_randomly(cur_points,cur_no_of_points,cur_error,cur_label)
 int compare(ptr1,ptr2) 
      struct unidim  *ptr1,*ptr2;
 {
-  float x;
+  double x;
   
   x = (*ptr1).value - (*ptr2).value;
   
@@ -427,12 +427,12 @@ int compare(ptr1,ptr2)
 /*		the "qsort" system call in this routine, and in the	*/
 /*		module "find_values".					*/
 /************************************************************************/
-float linear_split(no_of_eff_points) 
+double linear_split(no_of_eff_points)
      int no_of_eff_points;
 {
   int i,j,from,to,bestsplit;
-  float temp,impurity_1d;
-  float newval;
+  double temp,impurity_1d;
+  double newval;
   int l1,l2,r1,r2;
   int compare();
